refactor: update the url to sector-list
