+++ src/app/(main)/report/sector-list/_components/sector-row.tsx
